backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
django-cors-headers==4.7.0
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
kombu==5.5.4
packaging==25.0
Pillow==11.3.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pyjwt==2.10.1
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
